Fix cppcheck warnings with 1.63 on RHEL6, speedup execution with CMake subtree update
