+++ .gitexternals
@@ -1,2 +1,2 @@
 # -*- mode: cmake -*-
-# CMake/common https://github.com/Eyescale/CMake.git 182dd65
+# CMake/common https://github.com/Eyescale/CMake.git 19a6868
