+++ eq/fabric/subPixel.h
@@ -20,6 +20,7 @@
 #define EQFABRIC_SUBPIXEL_H
 
 #include <eq/fabric/api.h>
+#include <eq/fabric/types.h>
 #include <lunchbox/bitOperation.h>
 #include <lunchbox/log.h>
 #include <lunchbox/types.h>
@@ -28,7 +29,6 @@ namespace eq
 {
 namespace fabric
 {
-    class SubPixel;
     std::ostream& operator << ( std::ostream& os, const SubPixel& subPixel );
 
     /**
