+++ eq/server/types.h
@@ -37,18 +37,25 @@ class CompoundListener;
 class CompoundVisitor;
 class Config;
 class ConfigVisitor;
+class DFREqualizer;
 class Equalizer;
 class Frame;
 class FrameData;
+class FramerateEqualizer;
 class Layout;
+class LoadEqualizer;
+class MonitorEqualizer;
 class Node;
 class NodeFactory;
 class Observer;
 class Pipe;
 class Segment;
 class Server;
+class TileEqualizer;
 class TileQueue;
+class TreeEqualizer;
 class View;
+class ViewEqualizer;
 class Window;
 
 typedef std::vector< Config* >   Configs;
